Player moving cont.
Player can move but unity crashes when you try to move more than once
