+++ GameProject/Assets/Scenes/SampleScene.unity
@@ -325,6 +325,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   player: {fileID: 1542029588}
   rawGridPoint: {x: 0, y: 0, z: 0}
+  movableRadius: 8
   distanceY: 5
   offset: 1
   gridSize: 1
@@ -435,6 +436,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bbbbe76ca5c723e4d9ac6e8d0092fb59, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  pathToFollow: []
 --- !u!114 &1542029588
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -447,9 +449,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 82610a70f6281184b93a65d15522348b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _findPath: {fileID: 1542029587}
+  pathFinder: {fileID: 0}
   currentPos: {x: 0, y: 0, z: 0}
-  maxDistance: 10
 --- !u!224 &1542029590
 RectTransform:
   m_ObjectHideFlags: 0
